dev-0.3.5 新增任务报警设置 新增docker容器模式

+++ src/main/kotlin/dzt/studio/dppservice/schedule/FlinkJobsSchedules.kt
@@ -6,6 +6,7 @@ import dzt.studio.dppservice.dao.job.DppJobListDao
 import dzt.studio.dppservice.domain.job.DppJobList
 import dzt.studio.dppservice.domain.job.JobCheckPoint
 import dzt.studio.dppservice.domain.job.JobCurrentStatus
+import dzt.studio.dppservice.util.DingTalkRobotBuilder
 import dzt.studio.dppservice.util.JobStatus
 import org.slf4j.LoggerFactory
 import org.springframework.beans.factory.annotation.Autowired
@@ -13,6 +14,7 @@ import org.springframework.scheduling.annotation.EnableScheduling
 import org.springframework.scheduling.annotation.Scheduled
 import org.springframework.stereotype.Component
 import org.springframework.web.client.RestTemplate
+import java.util.*
 
 
 /**
@@ -46,7 +48,7 @@ class FlinkJobsSchedules {
                     it.containerUrl!! + "/jobs/{1}/checkpoints", Any::class.java, it.appId
                 )
                 val jobCheckPoint = JSON.parseObject(JSON.toJSONString(rc), JobCheckPoint::class.java)
-                if(!(it.jobStatus == "BUILDING" && jobCurrentStatus.state=="CANCELED")){
+                if (!(it.jobStatus == "BUILDING" && jobCurrentStatus.state == "CANCELED")) {
                     val dppJobList = DppJobList()
                     dppJobList.id = it.jobId
                     dppJobList.jobName = jobCurrentStatus.name
@@ -57,14 +59,26 @@ class FlinkJobsSchedules {
                     }
                     dppJobListDAO!!.updateByPrimaryKeySelective(dppJobList)
                 }
+                if (it.warningEnable) {
+                    when (jobCurrentStatus.state) {
+                        "FAILED" -> {
+                            val msg = "异常失败"
+                            buildWarn(it.jobName!!, it.warnTo!!, msg)
+                        }
+                        "RESTARTING" -> {
+                            val msg = "正在自动重启"
+                            buildWarn(it.jobName!!, it.warnTo!!, msg)
+                        }
+                    }
+                }
             } catch (e: Exception) {
                 if (!e.localizedMessage.contains("404 Not Found")) {
                     logger.error(e.localizedMessage)
-                    if(it.jobStatus == "FINISHED"){
+                    if (it.jobStatus == "FINISHED" && it.containerType == 2) {
                         it.containerId?.let { it1 -> dppContainerInfoDAO!!.deleteByContainerId(it1) }
                     }
                 }
-                if(it.jobStatus != "BUILDING"){
+                if (it.jobStatus != "BUILDING") {
                     val dppJobList = DppJobList()
                     dppJobList.id = it.jobId
                     dppJobList.jobStatus = JobStatus.FINISHED.toString()
@@ -73,4 +87,12 @@ class FlinkJobsSchedules {
             }
         }
     }
+
+    fun buildWarn(jobName: String, token: String, warnMsg: String) {
+        val warnTime = Date()
+        val msg = "\n 任务：$jobName \n 报警时间：$warnTime \n 报警信息：$warnMsg，请及时查看。"
+        DingTalkRobotBuilder.send(token, msg)
+    }
+
+
 }
